added legacy files for old stuff. removed test files from assets/ and fixed gitignore. moved back config_error/ to src/. Added transactions messages (attempted, validated, completed, combined) recording. function to load data from csv, with automatic retrival of correct folder based on the wanted metric (end of experiment or evolution), with post processing (mean, max, min, ...). multi box plot of performances, for any number of combinations, with possible a second quantity for a second y axis, and automatic x label creation and different boxplots for dishonests. Function for building automatically filenames lists for plotting. Dictionaries for bad and best behavior parameters combiantions. Function for querying if a filename belongs to that dictionaries. Pruning of bad/not best configurations when running experiments. Testing new combining stategies for behaviours. New strategies based on age and reputation weighting, and rework of running average reputation strategy.
