--- assets/template_new.json
@@ -1,78 +1,0 @@
-{
-  "width": 1200,
-  "height": 600,
-  "food": {
-    "x": 200,
-    "y": 300,
-    "radius": 50
-  },
-  "nest": {
-    "x": 1000,
-    "y": 300,
-    "radius": 50
-  },
-  "simulation_steps": 5000,
-  "simulation_seed": 5684436,
-  "number_runs": 6,
-  "visualization": {
-    "activate": false,
-    "fps": 60
-  },
-  "random_walk": {
-    "random_walk_factor": 0.9,
-    "levi_factor": 1.4
-  },
-  "agent": {
-    "radius": 8,
-    "speed": 2.5,
-    "communication_radius": 50,
-    "communication_stop_time": 0,
-    "communication_cooldown": 0,
-    "noise": {
-      "class": "BimodalNoise",
-      "parameters": {
-				"noise_sampling_mu": 0.05,
-				"noise_sampling_sigma": 0.05,
-				"noise_sd": 0.05
-		}
-    },
-    "fuel_cost": 0
-  },
-  "behaviors":[
-    {
-		"class": "ReputationRankingBehavior",
-		"population_size": 24,
-		"parameters": {
-			"ranking_threshold": 0.3
-			}
-		},
-		{
-		"class": "SaboteurReputationRankingBehavior",
-		"population_size": 1,
-		"parameters": {
-			"lie_angle": 90,
-			"ranking_threshold": 0.3
-			}
-		}
-  ],
-  "payment_system": {
-    "class": "DelayedPaymentPaymentSystem",
-    "initial_reward": 1,
-    "parameters": {
-      "information_share": 0.5
-    }
-  },
-  "market": {
-    "class": "FixedPriceMarket",
-    "parameters": {
-      "reward": 1
-    }
-  },
-  "data_collection": {
-    "output_directory": "/home/uga/ing/tesi/information-market/data/CLUSTER/ranking/",
-    "filename": "24r_NP_90LIA_0.3RT_0.05SMU_0.05SSD_0.05NSD.csv",
-    "metrics": ["transactions"],
-    "precise_recording_interval": 100,
-    "transactions_log": false
-      }
-}
