+++ src/config.py
@@ -6,7 +6,7 @@ LOG_COMPLETED_TRANSATIONS:bool = False
 LOG_REJECTED_TRANSATIONS:bool = False
 
 PROJECT_DIR="/home/uga/ing/tesi/information-market/"
-CONFIG_ERRORS_DIR=f"{PROJECT_DIR}config_errors/"
+CONFIG_ERRORS_DIR=f"{PROJECT_DIR}src/config_errors/"
 ERRORS_LOG_FILE=f"{PROJECT_DIR}src/error.log"
 
 DATA_FOLDER=f"{PROJECT_DIR}data/"
