+++ src/model/behavior.py
@@ -28,11 +28,16 @@ BEHAVIORS_NAME_DICT = {  "n": "Naive",
                         "Nn": "Naive",
                         "s": "Sceptical",
                         "Ns": "Sceptical",
-                        "r": "Reputation Ranking",
-                        "v": "Variable Scepticism",
-                        "Nv": "Variable Scepticism",
-                        "t": "Reputation Threshold",
-                        "w": "Reputation Weighted",
+                        # "r": "Reputation Ranking",
+                        "r": "Rep. Ranking",
+                        # "v": "Variable Scepticism",
+                        # "Nv": "Variable Scepticism",
+                        "v": "Variable Sc.",
+                        "Nv": "Variable Sc.",
+                        "t": "Rep. Threshold",
+                        "w": "Rep. Weighted",
+                        # "t": "Reputation Threshold",
+                        # "w": "Reputation Weighted",
                     }
 SUB_FOLDERS_DICT={  "n": "naive",
                     "Nn": "new_naive",
@@ -44,7 +49,8 @@ SUB_FOLDERS_DICT={  "n": "naive",
                     "t": "wealth_threshold",
                     "w": "wealth_weighted",
                     }
-PARAMS_NAME_DICT={  "ST": "scepticims threshold",
+PARAMS_NAME_DICT={
+                    "ST": "scepticism threshold",
                     "RT": "ranking threshold",
                     "CM": "comparison method",
                     "SC": "scaling",
@@ -58,10 +64,7 @@ PARAMS_NAME_DICT={  "ST": "scepticims threshold",
                     "NMU": "uniform noise mean",
                     "NRANG": "uniform noise range",
                     "SAB": "saboteur performance",
-                }
-NOISE_PARAMS_DICT={ "bimodal": ["SMU","SSD","NSD"],
-                    "uniform": ["NMU","NRANG","SAB"],
-                }
+                    }
 BEHAVIOR_PARAMS_DICT = {"n": [],
                         "Nn": [],
                         "s": ["ST"],
@@ -72,22 +75,54 @@ BEHAVIOR_PARAMS_DICT = {"n": [],
                         "t": ["CM","SC"],
                         "w": [],
                         }
-BAD_COMBINATIONS={
+NOISE_PARAMS_DICT={ "bimodal": ["SMU","SSD","NSD"],
+                    "uniform": ["NMU","NRANG","SAB"],
+                }
+BEST_PARAM_COMBINATIONS_DICT={
+                            "n": [
+                                ['NP',[]],
+                                ['P',[]],
+                                ],
+                            "Nn": [['NP',[]],['P',[]],],
+                            "s": [['NP',['025']],['P',['025']],],
+                            "Ns": [
+                                ['NP',['025']],
+                                ['P',['025']],
+                                ],
+                            "r": [
+                                ['P', ['03']],
+                                ],
+                            "v": [],
+                            "Nv": [
+                                ['P', ['allavg', '03', '025', 'exponential']],
+                                ['P', ['allavg', '03', '025', 'ratio']],
+                                ['NP', ['allmax', '03', '025', 'ratio']],
+                                ],
+                            "t": [
+                                ['P', ['allavg', '03']],
+                                ['P', ['allavg', '05']],
+                                ],
+                            "w": [],
+                            }
+
+BAD_PARAM_COMBINATIONS_DICT={
                     "n": [],
                     "Nn": [],
                     "s": [],
                     "Ns": [],
                     "r": [
-                        # ['', ['']],
+                        ['NP', ['03']],
+                        ['NP', ['05']],
                         ],
                     "v": [],
                     "Nv": [
                         ['NP', ['allavg', '03', '025', 'exponential']],
+                        ['NP', ['allavg', '03', '025', 'ratio']],
                         ['NP', ['allavg', '05', '025', 'exponential']],
                         ['NP', ['allavg', '08', '025', 'exponential']],
                         ['NP', ['allmax', '03', '025', 'exponential']],
                         ['NP', ['allmax', '08', '025', 'exponential']],
-                        ['NP', ['allavg', '08', '025', 'exponential']],
+                        ['NP', ['allmax', '08', '025', 'exponential']],
                         ['P', ['allavg', '08', '025', 'exponential']],
                         ['P', ['allavg', '08', '025', 'ratio']],
                         ['P', ['allmax', '03', '025', 'exponential']],
@@ -103,6 +138,7 @@ BAD_COMBINATIONS={
                         ['NP', ['allavg', '08']],
                         ['NP', ['allmax', '03']],
                         ['NP', ['allmax', '05']],
+                        ['NP', ['allmax', '08']],
                         ['P', ['allmax', '05']],
                         ['P', ['allmax', '08']],
                         ],
@@ -131,7 +167,10 @@ BAD_COMBINATIONS={
                     #     ['P', ['allmax', '0.5']],
                     #     ['P', ['allmax', '0.8']],
                     #     ],\\
-                    "w": [],
+                    "w": [
+                        ['NP',[]],
+                        ['P',[]],
+                    ],
                 }
 
 
@@ -413,8 +452,9 @@ class TemplateBehaviour(Behavior):
         self.navigation_table.rotate_from_angle(-get_orientation_from_vector(self.dr))
 
 
+
 #################################################################################################
-## BASE BEHAVIORS
+## NAIVE BEHAVIORS
 class NaiveBehavior(Behavior):
     def __init__(self):
         super().__init__()
@@ -542,66 +582,6 @@ class NaiveBehavior(Behavior):
         self.navigation_table.rotate_from_angle(-get_orientation_from_vector(self.dr))
 
 
-#####################################
-# #
-class CarefulBehavior(NaiveBehavior):
-    def __init__(self, security_level=3):
-        super(CarefulBehavior, self).__init__()
-        self.color = "deep sky blue"
-        self.security_level = security_level
-        self.pending_information = {location: {} for location in Location}
-
-
-    def buy_info(self, session: CommunicationSession):
-        for location in Location:
-            metadata = session.get_metadata(location)
-            metadata_sorted_by_age = sorted(metadata.items(), key=lambda item: item[1]["age"])
-            for bot_id, data in metadata_sorted_by_age:
-                if data["age"] < self.navigation_table.get_age_for_location(location) and bot_id not in \
-                        self.pending_information[
-                            location]:
-                    try:
-                        other_target = session.make_transaction(neighbor_id=bot_id, location=location)
-                        other_target.set_distance(other_target.get_distance() + session.get_distance_from(bot_id))
-                        if not self.navigation_table.is_information_valid_for_location(location):
-                            self.navigation_table.replace_information_entry(location, other_target)
-                        else:
-                            self.pending_information[location][bot_id] = other_target
-                            if len(self.pending_information[location]) >= self.security_level:
-                                self.combine_pending_information(location)
-                    except InsufficientFundsException:
-                        pass
-                    except NoInformationSoldException:
-                        pass
-
-
-    def combine_pending_information(self, location):
-        distances = [t.get_distance() for t in self.pending_information[location].values()]
-        mean_distance = np.mean(distances, axis=0)
-        best_target = min(self.pending_information[location].values(),
-                          key=lambda t: norm(t.get_distance() - mean_distance))
-        self.navigation_table.replace_information_entry(location, best_target)
-        self.pending_information[location].clear()
-
-
-    def step(self, api):
-        super().step(api)
-        self.update_pending_information()
-
-
-    def update_pending_information(self):
-        for location in Location:
-            for target in self.pending_information[location].values():
-                target.update(self.dr)
-                target.rotate(-get_orientation_from_vector(self.dr))
-
-
-    def debug_text(self):
-        return f"size of pending: {[len(self.pending_information[l]) for l in Location]}\n" \
-               f"{self.pending_information[Location.FOOD]}\n" \
-               f"{self.pending_information[Location.NEST]}"
-
-
 class SaboteurBehavior(NaiveBehavior):
     def __init__(self, lie_angle=90):
         super().__init__()
@@ -611,17 +591,6 @@ class SaboteurBehavior(NaiveBehavior):
     def sell_info(self, location):
         t = copy.deepcopy(self.navigation_table.get_information_entry(location))
         t.rotate(self.lie_angle)
-        return t
-
-
-class GreedyBehavior(NaiveBehavior):
-    def __init__(self):
-        super().__init__()
-        self.color = "green"
-
-    def sell_info(self, location):
-        t = copy.deepcopy(self.navigation_table.get_information_entry(location))
-        t.age = 1
         return t
 
 
@@ -696,15 +665,6 @@ class ScepticalBehavior(NaiveBehavior):
                 target.rotate(-get_orientation_from_vector(self.dr))
 
 
-class FreeRiderBehavior(ScepticalBehavior):
-    def __init__(self):
-        super().__init__()
-        self.color = "pink"
-
-    def sell_info(self, location):
-        return None
-
-
 class ScaboteurBehavior(ScepticalBehavior):
     def __init__(self, lie_angle=90, threshold=.25):
         super().__init__()
@@ -718,285 +678,9 @@ class ScaboteurBehavior(ScepticalBehavior):
         return t
 
 
-class ScepticalGreedyBehavior(ScepticalBehavior):
-    def __init__(self):
-        super().__init__()
-        self.color = "green"
-
-    def sell_info(self, location):
-        t = copy.deepcopy(self.navigation_table.get_information_entry(location))
-        t.age = 1
-        return t
-
-
-###################################################################################
-# BEHAVIOURS WITH REPUTATION (SYSTEMIC) PROTECTION ################################
-
-# class SimpleReputationBehaviour
-
-
-
-class ReputationWealthBehaviour(NaiveBehavior):
-    def __init__(self):
-        super(ReputationWealthBehaviour, self).__init__()
-        #TODO if reputation level raises, has pending info sense?
-        # self.pending_information = {location: {} for location in Location}
-        self.required_information=RequiredInformation.GLOBAL
-
-
-    @abstractmethod
-    def verify_reputation(self,payment_database:PaymentDB, session: CommunicationSession, bot_id):
-        pass
-
-    def get_reputation_score(self, payment_database:PaymentDB, bot_id):
-        return payment_database.get_reward(bot_id)
-
-    def buy_info(self, session: CommunicationSession, payment_database: PaymentDB):
-        for location in Location:
-            metadata = session.get_metadata(location)
-            metadata_sorted_by_age = sorted(metadata.items(), key=lambda item: item[1]["age"])
-            for bot_id, data in metadata_sorted_by_age:
-                if data["age"] < self.navigation_table.get_age_for_location(location):
-                    #  and bot_id not in self.pending_information[location]:
-                    try:
-                        other_target = session.make_transaction(neighbor_id=bot_id, location=location)
-                        other_target.set_distance(other_target.get_distance() + session.get_distance_from(
-                            bot_id))
-
-                        if not self.navigation_table.is_information_valid_for_location(location) or \
-                                self.verify_reputation(payment_database,session, bot_id):
-                            new_target = self.strategy.combine(self.navigation_table.get_information_entry(location),
-                                                               other_target,
-                                                               np.array([0, 0]))
-                            self.navigation_table.replace_information_entry(location, new_target)
-                            # self.pending_information[location].clear()
-                    except InsufficientFundsException:
-                        pass
-                    except NoInformationSoldException:
-                        pass
-    def step(self, api):
-        return super().step(api)
-
-
-class ReputationTresholdBehaviour(ReputationWealthBehaviour):
-    def __init__(self):
-        super(ReputationTresholdBehaviour, self).__init__()
-        #TODO if reputation level raises, has pending info sense?
-        # self.pending_information = {location: {} for location in Location}
-
-    @abstractmethod
-    #TODO
-    # def get_threshold_value(self, session: CommunicationSession):
-    def get_threshold_value(self):
-        pass
-
-    def verify_reputation(self,payment_database:PaymentDB, session: CommunicationSession, bot_id):
-        # return self.get_reputation_score(session, bot_id) >\
-        return self.get_reputation_score(payment_database, bot_id) >\
-                 self.get_threshold_value(payment_database, session)
-
-
-class ReputationStaticThresholdBehavior(ReputationTresholdBehaviour):
-    def __init__(self,threshold=1):
-        super(ReputationStaticThresholdBehavior, self).__init__()
-        self.reputation_threshold=threshold
-
-    def get_threshold_value(self, payment_database:PaymentDB,session: CommunicationSession):
-            return self.reputation_threshold
-
-
-class SaboteurReputationStaticThresholdBehavior(ReputationStaticThresholdBehavior):
-    def __init__(self, threshold=1,lie_angle=90):
-        super().__init__(threshold)
-        self.color = "red"
-        self.lie_angle = lie_angle
-
-    def sell_info(self, location):
-        t = copy.deepcopy(self.navigation_table.get_information_entry(location))
-        t.rotate(self.lie_angle)
-        return t
-
-
-class ReputationDynamicThresholdBehavior(ReputationTresholdBehaviour):
-    def __init__(self,comparison_method="allmax",scaling=1):
-        super(ReputationDynamicThresholdBehavior, self).__init__()
-        self.comparison_method=comparison_method
-        self.scaling=scaling
-
-    def get_threshold_value(self, payment_database:PaymentDB,session: CommunicationSession):
-        """
-        allmax: selects only above a certain percentage of maximum wealth (wealthiest bots), of all robots
-        allavg:selects only above certain percentage of average wealth, considering all robots
-        allmin: selects only above a certain percentage of minimum wealth (poorest bots), of all robots
-        TODO: all_rise:  selects above a certaint value, increasing with time, starting from a certain level, of all robots
-
-        DEPRECATED:
-        neighmax: selects only above a certain percentage of maximum wealth (wealthiest bots), of neighbors
-        neighavg: selects only above certain percentage of average wealth, considering neighbors
-        neighmin: selects only above a certain percentage of minimum wealth (poorest bots), of neighbors
-        """
-        # extension, metric=re.split("_",self.comparison_method)
-        extension, metric=self.comparison_method[:-3],self.comparison_method[-3:]
-        reputation_dict = {"all":{
-                                "max":payment_database.get_highest_reward,
-                                 "avg":payment_database.get_average_reward,
-                                 "min":payment_database.get_lowest_reward,
-                                 },
-                            "neigh":{
-                                    "max":session.get_max_neighboor_reward,
-                                    "avg":session.get_average_neighbor_reward,
-                                    "min":session.get_min_neighboor_reward,
-                                }
-                            }
-        try:
-            return self.scaling*reputation_dict[extension][metric]()
-        except KeyError:
-            exit(1)
-            # return super().reputation_threshold(session, self.method)
-
-
-
-class SaboteurReputationDynamicThresholdBehavior(ReputationDynamicThresholdBehavior):
-    def __init__(self, comparison_method="allmax",scaling=1,lie_angle=90):
-        super().__init__(comparison_method,scaling)
-        self.color = "red"
-        self.lie_angle = lie_angle
-
-    def sell_info(self, location):
-        t = copy.deepcopy(self.navigation_table.get_information_entry(location))
-        t.rotate(self.lie_angle)
-        return t
-
-
-#########################################################################################
-# SKEPTICISM + REPUTATION WITH WEALTH BEHAVIOURS
-
-class ScepticalReputationBehavior(NaiveBehavior):
-    def __init__(self,comparison_method="allavg",scaling=1,scepticism_threshold=.25,weight_method="linear"):
-        super(ScepticalReputationBehavior, self).__init__()
-        self.scepticism_threshold=scepticism_threshold
-        self.weight_method=weight_method
-        self.comparison_method=comparison_method
-        self.scaling=scaling
-        self.pending_information = {location: {} for location in Location}
-        self.required_information=RequiredInformation.GLOBAL
-
-    @staticmethod
-    def difference_score(current_vector, bought_vector):
-        v_norm = norm(current_vector)
-        score = norm(current_vector - bought_vector) / v_norm if v_norm > 0.0 else 1000
-        return score
-
-
-    def step(self, api):
-        super().step(api)
-        self.update_pending_information()
-
-
-    def update_pending_information(self):
-        for location in Location:
-            for target in self.pending_information[location].values():
-                target.update(self.dr)
-                target.rotate(-get_orientation_from_vector(self.dr))
-
-
-    def get_reputation_score(self,payment_database:PaymentDB,bot_id):
-        return payment_database.get_reward(bot_id)
-
-
-    def get_skepticism_threshold(self,payment_database:PaymentDB,bot_id):
-        reputation_score=self.get_reputation_score(payment_database,bot_id)
-        # extension, metric=re.split("_",self.comparison_method)
-        extension, metric=self.comparison_method[:-3],self.comparison_method[-3:]
-        reputation_dict = {"all":{
-                            "max":payment_database.get_highest_reward,
-                            "avg":payment_database.get_average_reward,
-                            "min":payment_database.get_lowest_reward,
-                            },
-                        }
-        try:
-            # return self.weighted_sceticism(reputation_score,reputation_dict[extension][metric])
-            return self.weight_scepticism(reputation_score,reputation_dict[extension][metric]())
-        except KeyError:
-            return self.scepticism_threshold
-
-
-    def weight_scepticism(self,reputation_score,metric_score):
-        #TODO fix wrong initialization of behaviour, causing some to have default
-        # print(self.weight_method)
-        try:
-            if self.weight_method=="linear":
-                if reputation_score>metric_score:
-                    threshold= self.scepticism_threshold-reputation_score/metric_score
-                else:
-                    threshold= self.scepticism_threshold+reputation_score/metric_score
-                return max(0,threshold)
-            elif self.weight_method=="ratio":
-                return self.scepticism_threshold*reputation_score/metric_score
-            elif self.weight_method=="exponential":
-                return self.scepticism_threshold*(reputation_score/metric_score)**2
-            elif self.weight_method=="logarithmic":
-                return self.scepticism_threshold*np.log(reputation_score/metric_score)
-        except:
-            return self.scepticism_threshold
-
-
-    def buy_info(self, payment_database:PaymentDB,session: CommunicationSession):
-        for location in Location:
-            metadata = session.get_metadata(location)
-            metadata_sorted_by_age = sorted(metadata.items(), key=lambda item: item[1]["age"])
-            for bot_id, data in metadata_sorted_by_age:
-                if data["age"] < self.navigation_table.get_age_for_location(location) and bot_id not in \
-                        self.pending_information[location]:
-                    try:
-                        other_target = session.make_transaction(neighbor_id=bot_id, location=location)
-                        other_target.set_distance(other_target.get_distance() + session.get_distance_from(
-                            bot_id))
-                        if not self.navigation_table.is_information_valid_for_location(location) or \
-                                self.difference_score(
-                                    self.navigation_table.get_relative_position_for_location(location),
-                                    other_target.get_distance())\
-                                < self.get_skepticism_threshold(payment_database,bot_id):
-                            new_target = self.strategy.combine(self.navigation_table.get_information_entry(location),
-                                                               other_target,
-                                                               np.array([0, 0]))
-                            self.navigation_table.replace_information_entry(location, new_target)
-                            self.pending_information[location].clear()
-                        else:
-                            for target in self.pending_information[location].values():
-                                if self.difference_score(target.get_distance(),
-                                                         other_target.get_distance()) \
-                                        < self.get_skepticism_threshold(payment_database,bot_id):
-                                    new_target = self.strategy.combine(target,
-                                                                       other_target,
-                                                                       np.array([0, 0]))
-                                    self.navigation_table.replace_information_entry(location, new_target)
-                                    self.pending_information[location].clear()
-                                    break
-                            else:
-                                self.pending_information[location][bot_id] = other_target
-                    except InsufficientFundsException:
-                        pass
-                    except NoInformationSoldException:
-                        pass
-
-
-class SaboteurScepticalReputationBehavior(ScepticalReputationBehavior):
-    def __init__(self,comparison_method="allavg",scaling=1,scepticism_threshold=.25,weight_method="logarithmic", lie_angle=90):
-        super().__init__(comparison_method,scaling,scepticism_threshold,weight_method)
-        self.color = "red"
-        self.lie_angle = lie_angle
-
-    def sell_info(self, location):
-        t = copy.deepcopy(self.navigation_table.get_information_entry(location))
-        t.rotate(self.lie_angle)
-        return t
-
-
-######################################################
-# BASED ON NEW TEMPLATE
-
-#TESTED
+##################################################################################################
+##################################################################################################
+# # NEW CLASSICAL BEHAVIOURS
 class NewNaiveBehavior(TemplateBehaviour):
     def __init__(self):
         super().__init__()
@@ -1021,7 +705,6 @@ class NewSaboteurBehavior(NewNaiveBehavior):
         return t
 
 
-#TESTED
 class NewScepticalBehavior(TemplateBehaviour):
     def __init__(self,scepticism_threshold=.25):
         super().__init__()
@@ -1047,7 +730,6 @@ class NewScepticalBehavior(TemplateBehaviour):
         v_norm = norm(current_vector)
         score = norm(current_vector - bought_vector) / v_norm if v_norm > 0 else 1000
         return score
-
     #TODO test which is better
     # def difference_score(self,location,other_target:Target):
     #     current_vector=self.navigation_table.get_relative_position_for_location(location)
@@ -1099,7 +781,6 @@ class NewScepticalBehavior(TemplateBehaviour):
         return False
 
 
-#TESTED
 class NewScaboteurBehavior(NewScepticalBehavior):
     def __init__(self,lie_angle=90,scepticism_threshold=.25):
         super().__init__(scepticism_threshold)
@@ -1112,7 +793,10 @@ class NewScaboteurBehavior(NewScepticalBehavior):
         return t
 
 
-#TO TEST
+##################################################################################################
+#################################################################
+# BEHAVIOURS WITH REPUTATION (SYSTEMIC PROTECTION)
+
 class WealthWeightedBehavior(TemplateBehaviour):
     def __init__(self):
         super().__init__()
@@ -1156,9 +840,11 @@ class ReputationRankingBehavior(TemplateBehaviour):
     def __init__(self,ranking_threshold=.5):
         super().__init__()
         self.required_information=RequiredInformation.GLOBAL
-        self.information_ordering_metric="reputation"
-        self.strategy=strategy_factory("WeightedAverageAgeStrategy")
-        self.RANKING_THRESHOLD=ranking_threshold
+        # self.information_ordering_metric="reputation"
+        self.information_ordering_metric="age"
+        # self.strategy=strategy_factory("WeightedAverageAgeStrategy")
+        self.strategy=strategy_factory("RunningWeightedAverageReputationStrategy")
+        self.ranking_threshold=ranking_threshold
 
 
     def test_data_validity(self, location: Location, data,_,__):
@@ -1178,11 +864,7 @@ class ReputationRankingBehavior(TemplateBehaviour):
     def compute_seller_percentile(self,payment_database:PaymentDB,seller_id):
         rank=payment_database.get_wallet_ranking(seller_id)
         percentile=1-rank/len(payment_database.database)
-        return percentile >= self.RANKING_THRESHOLD
-
-    #TODO
-    # def stop_buying_process(self):
-    #     return False
+        return percentile >= self.ranking_threshold
 
 
 class SaboteurReputationRankingBehavior(ReputationRankingBehavior):
@@ -1202,7 +884,8 @@ class WealthThresholdBehavior(TemplateBehaviour):
         super().__init__()
         self.required_information=RequiredInformation.GLOBAL
         self.information_ordering_metric="age"
-        self.strategy=strategy_factory("WeightedAverageAgeStrategy")
+        # self.strategy=strategy_factory("WeightedAverageAgeStrategy")
+        self.strategy=strategy_factory("RunningWeightedAverageReputationStrategy")
         self.scaling=scaling
         self.comparison_method=comparison_method
 
@@ -1277,13 +960,15 @@ class NewScepticalReputationBehavior(NewScepticalBehavior):
         super().__init__(scepticism_threshold)
         self.required_information=RequiredInformation.GLOBAL
         self.information_ordering_metric="age"
-        self.strategy=strategy_factory("WeightedAverageAgeStrategy")
+        # self.strategy=strategy_factory("WeightedAverageAgeStrategy")
+        self.strategy=strategy_factory("RunningWeightedAverageReputationStrategy")
         self.scaling=scaling
         self.comparison_method=comparison_method
         self.weight_method=weight_method
 
     def get_scepticism_threshold(self,payment_database:PaymentDB,seller_id):
         reputation_score=payment_database.get_reward(seller_id)
+        #TODO could substitute with "-"
         # extension, metric=re.split("_",self.comparison_method)
         extension, metric=self.comparison_method[:-3],self.comparison_method[-3:]
 
@@ -1295,15 +980,12 @@ class NewScepticalReputationBehavior(NewScepticalBehavior):
                             },
                         }
         try:
-            # return self.weighted_sceticism(reputation_score,reputation_dict[extension][metric])
             return self.weight_scepticism(reputation_score,reputation_dict[extension][metric]())
         except KeyError:
             return self.scepticism_threshold
 
 
     def weight_scepticism(self,reputation_score,metric_score):
-        #TODO fix wrong initialization of behaviour, causing some to have default
-        # print(self.weight_method)
         try:
             if self.weight_method=="ratio":
                 return self.scepticism_threshold*reputation_score/(metric_score*self.scaling)
@@ -1331,3 +1013,4 @@ class NewSaboteurScepticalReputationBehavior(NewScepticalReputationBehavior):
         t = copy.deepcopy(self.navigation_table.get_information_entry(location))
         t.rotate(self.lie_angle)
         return t
+
